[BBC-9075] Migrate `DatePicker` to MDX
